working calendar, updates to calender mycalendar and events
