+++ package.json
@@ -10,7 +10,6 @@
     "@testing-library/user-event": "^12.7.1",
     "axios": "^0.21.1",
     "bcryptjs": "^2.4.3",
-    "date-fns": "^2.17.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "express-session": "^1.17.1",
