+++ src/Components/Calendar/Calendar.js
@@ -1,119 +1,60 @@
-import React from "react";
-import dateFns from "date-fns";
+import React, {Component} from 'react' 
+import FullCalendar from '@fullcalendar/react'
+import dayGridPlugin from '@fullcalendar/daygrid'
+// import timeGridPlugin from '@fullcalendar/timegrid'
+import interactionPlugin from '@fullcalendar/interaction'
+import "@fullcalendar/core/main.css";
+// import "@fullcalendar/timegrid/main.css"
 
-class Calendar extends React.Component {
-  state = {
-    currentMonth: new Date(),
-    selectedDate: new Date()
-  };
+const events =[{title: '', date: new Date()}]
 
-  renderHeader() {
-    const dateFormat = "MMMM YYYY";
+class Calendar extends Component{
 
-    return (
-      <div className="header row flex-middle">
-        <div className="col col-start">
-          <div className="icon" onClick={this.prevMonth}>
-            chevron_left
-          </div>
-        </div>
-        <div className="col col-center">
-          <span>{dateFns.format(this.state.currentMonth, dateFormat)}</span>
-        </div>
-        <div className="col col-end" onClick={this.nextMonth}>
-          <div className="icon">chevron_right</div>
-        </div>
-      </div>
-    );
+formatEvents =() => {
+    return this.props.appointments.map(appointment => {
+              const {title, end, start} = appointment
+  
+              let startTime = new Date(start)
+              let endTime = new Date(end)
+  
+              return {
+                title, 
+                start: startTime,
+                end: endTime, 
+                extendedProps: {...appointment}
+              }
+          })
   }
 
-  renderDays() {
-    const dateFormat = "dddd";
-    const days = [];
-
-    let startDate = dateFns.startOfWeek(this.state.currentMonth);
-
-    for (let i = 0; i < 7; i++) {
-      days.push(
-        <div className="col col-center" key={i}>
-          {dateFns.format(dateFns.addDays(startDate, i), dateFormat)}
-        </div>
-      );
-    }
-
-    return <div className="days row">{days}</div>;
-  }
-
-  renderCells() {
-    const { currentMonth, selectedDate } = this.state;
-    const monthStart = dateFns.startOfMonth(currentMonth);
-    const monthEnd = dateFns.endOfMonth(monthStart);
-    const startDate = dateFns.startOfWeek(monthStart);
-    const endDate = dateFns.endOfWeek(monthEnd);
-
-    const dateFormat = "D";
-    const rows = [];
-
-    let days = [];
-    let day = startDate;
-    let formattedDate = "";
-
-    while (day <= endDate) {
-      for (let i = 0; i < 7; i++) {
-        formattedDate = dateFns.format(day, dateFormat);
-        const cloneDay = day;
-        days.push(
-          <div
-            className={`col cell ${
-              !dateFns.isSameMonth(day, monthStart)
-                ? "disabled"
-                : dateFns.isSameDay(day, selectedDate) ? "selected" : ""
-            }`}
-            key={day}
-            onClick={() => this.onDateClick(dateFns.parse(cloneDay))}
-          >
-            <span className="number">{formattedDate}</span>
-            <span className="bg">{formattedDate}</span>
-          </div>
-        );
-        day = dateFns.addDays(day, 1);
-      }
-      rows.push(
-        <div className="row" key={day}>
-          {days}
-        </div>
-      );
-      days = [];
-    }
-    return <div className="body">{rows}</div>;
-  }
-
-  onDateClick = day => {
-    this.setState({
-      selectedDate: day
-    });
-  };
-
-  nextMonth = () => {
-    this.setState({
-      currentMonth: dateFns.addMonths(this.state.currentMonth, 1)
-    });
-  };
-
-  prevMonth = () => {
-    this.setState({
-      currentMonth: dateFns.subMonths(this.state.currentMonth, 1)
-    });
-  };
-
-  render() {
-    return (
-      <div className="calendar">
-        {this.renderHeader()}
-        {this.renderDays()}
-        {this.renderCells()}
-      </div>
-    );
-  }
+handleEventClick= ({event}) => {
+   this.props.openAppointment(event.extendedProps)
 }
-export default Calendar;
+
+handleEventDrop = (info) => {
+        if(window.confirm("Are you sure you want to change the event date?")){
+            console.log('change confirmed')
+
+            this.props.updateAppointment({...info.event.extendedProps, start: info.event.start, end: info.event.end})
+
+        } else {
+            console.log('change aborted')
+        }
+   }
+
+
+   render(){
+    return(
+        <div>
+            <FullCalendar
+            defaultView="dayGridMonth" 
+            plugins={[dayGridPlugin, interactionPlugin]}
+            editable={true}
+            eventDrop={this.handleEventDrop}
+            eventClick={this.handleEventClick}
+            events={this.formatEvents()}
+            />
+        </div>
+    )
+   }
+  }
+  export default Calendar
