+++ src/routes.js
@@ -5,7 +5,7 @@ import Blog from "./Components/Blog/Blog";
 import Home from "./Components/Home/Home";
 import Chat from "./Components/Chat/Chat";
 import Message from "./Components/Chat/Message";
-import Event from "./Components/Event/Events";
+import CalEvent from './Components/Event/Event'
 import Group from "./Components/Group/Group";
 import Profile from "./Components/Profile/Profile";
 
@@ -14,7 +14,7 @@ export default (
     <Route exact path="/" component={Home} />
     <Route path="/auth" component={Auth}/>
     <Route path="/blog" component={Blog} />
-    <Route path="/events" component={Event} />
+    <Route path="/cal-events" component={CalEvent}/>
     <Route path="/chat" component={Chat} />
     <Route path="/message" component={Message} />
     <Route path="/groups" component={Group} />
